+++ src/__tests__/attributeExtractorTest.ts
@@ -1,22 +1,23 @@
 import { removeWhites } from 'misc-utils-of-mine-generic'
-import Project, { ScriptTarget } from 'ts-morph'
 import { replaceFileFunctionCall } from '../replaceFileFunctionCall'
 import { defaultOptions } from '../replaceProjectFunctionCall'
 import { Attribute } from '../extractors/core/attribute'
-import { evaluate, Fn, evaluateAndError } from '../util'
+import { evaluateExtractorTestCode } from './testUtil'
+import Project from 'ts-morph'
 
 describe('AttributeExtractor', () => {
   it('dummy', () => {
     expect(1).toBe(1)
   })
 
-  it('should get and set strings asArgument', () => {
+  it('should get and set strings asArgument without errors', () => {
     const project = new Project()
     project.createSourceFile(
       'test.ts',
       `
 class C{}
 Attribute({target: C, name: 'attr1', value: 'hello'})
+const v = Attribute({target: C, name: 'attr1'})
     `
     )
     replaceFileFunctionCall(project.getSourceFile('test.ts')!, {
@@ -28,7 +29,8 @@ Attribute({target: C, name: 'attr1', value: 'hello'})
     expect(removeWhites(project.getSourceFile('test.ts')!.getText())).toContain(
       removeWhites(`
 class C{}
-Attribute({target: C, name: 'attr1', value: 'hello'}, {name: "attr1", node: 'hello'})
+Attribute({target: C, name: 'attr1', value: 'hello'}, "")
+const v = Attribute({target: C, name: 'attr1'}, Object.values(fileVariables).find(v=>v.extractorName==='Attribute'&&v.name==="attr1"))
     `)
     )
   })
@@ -40,6 +42,7 @@ Attribute({target: C, name: 'attr1', value: 'hello'}, {name: "attr1", node: 'hel
       `
  class C{}
  Attribute({target: C, name: 'attr1', value: 'hello'})
+ const v =  Attribute({target: C, name: 'attr1'})
      `
     )
     replaceFileFunctionCall(project.getSourceFile('test.ts')!, {
@@ -47,29 +50,22 @@ Attribute({target: C, name: 'attr1', value: 'hello'}, {name: "attr1", node: 'hel
       ...{ extractorDataMode: 'prependVariable' },
       project
     })
+    // console.log(project.getSourceFile('test.ts')!.getText());
     expect(removeWhites(project.getSourceFile('test.ts')!.getText())).toContain(
       removeWhites(`
-const __extractor_prepend__ = [{name: "attr1", node: 'hello'}]
+const __extractor_prepend__ = ["", ""]
+const fileVariables: {[name:string]: any} = {
+    "0_attr1_0": {value: 'hello', name: "attr1", index: 0, extractorName: "Attribute"}
+}
 class C{}
-Attribute({target: C, name: 'attr1', value: 'hello'}, __extractor_prepend__[0])
+Attribute({target: C, name: 'attr1', value: 'hello'}, "")
+const v =  Attribute({target: C, name: 'attr1'}, Object.values(fileVariables).find(v=>v.extractorName==='Attribute'&&v.name==="attr1"))
      `)
     )
   })
 
   it('emit execute extractor and evaluate here instead of spawning', () => {
-    //  const f = project.createSourceFile(
-
-    // )
-
     let r: any
-    //     let r = evaluateExtractorTestCode(`
-    // function test(){
-    //   class C{}
-    //   return Attribute({target: C, name: 'attr1', value: function(q){return q+2}})
-    // }
-    // `, 'Attribute', Attribute)
-    //     console.log(r.resultBefore, r.resultAfter, r.jsBefore, r.jsAfter);
-    //     console.log(r.jsAfter);
 
     r = evaluateExtractorTestCode(
       `
@@ -83,26 +79,8 @@ function test(){
       'Attribute',
       Attribute
     )
-    // console.log(r.resultBefore, r.resultAfter, r.jsBefore, r.jsAfter);
-    // console.log(r.jsAfter);
 
     expect(r.resultAfter).toBe(6)
-
-    // r= evaluateExtractorTestCode(`
-    // function test(){
-    //   class C{}
-    //   return Attribute({target: C, name: 'attr1'})
-    // }
-    // `, 'Attribute', Attribute)
-    // console.log(r);
-
-    // console.log('EVAL: ',  eval(js), 'END');
-
-    //      expect(removeWhites(project.getSourceFile('test.ts')!.getText())).toContain(removeWhites(`
-    // const __extractor_prepend__ = [{name: "attr1", node: 'hello'}]
-    // class C{}
-    // Attribute({target: C, name: 'attr1', value: 'hello'}, __extractor_prepend__[0])
-    //      `))
   })
 
   xit('should get and set nodes', () => {})
@@ -110,37 +88,3 @@ function test(){
 
   xit('should ?? with other value types (object, function, etc', () => {})
 })
-function evaluateExtractorTestCode(code: string, extractorName: string, extractorFn: Fn) {
-  const project = new Project({ compilerOptions: { target: ScriptTarget.ESNext } })
-  project.createSourceFile('test.ts', code)
-  const jsBefore = `
-(function(){
-${extractorFn.toString()}
-${project.emitToMemory().getFiles()[0].text}
-return test()
-  })()
-`.trim()
-
-  const resultBefore = evaluateAndError(jsBefore)
-  replaceFileFunctionCall(project.getSourceFile('test.ts')!, {
-    ...defaultOptions,
-    ...{ extractorDataMode: 'prependVariable' },
-    project
-  })
-  const jsAfter = `
-(function(){
-${extractorFn.toString()} 
-${project.emitToMemory().getFiles()[0].text}
-return test()
-  })()
-`.trim()
-  const resultAfter = evaluateAndError(jsAfter)
-
-  return {
-    resultBefore,
-    resultAfter,
-    jsBefore: jsBefore.replace(/\\n/gm, '\n'),
-    jsAfter: jsAfter.replace(/\\n/gm, '\n'),
-    project
-  } //, file:  project.getSourceFile('test.ts')!}
-}