+++ src/extractors/core/attribute.ts
@@ -13,20 +13,21 @@ import { quote } from 'misc-utils-of-mine-generic'
 
 /**
  * get/set names to nodes/types. Query nodes, like in CSS
-
-```ts
-// setting a type or node's attribute value
-
-Attribute<SomeInterface>({name: 'id', value: 'org.foo.Super'})
-Attribute({target: anImpl1, name: 'class', value: 'logger'})
-Attribute({target: aNode, name: 'otherNode', value: aFunctionDeclaration})
-
-// getting (previously setted) attribute values
-
-// if target is not passed at all the the variable won't be bound to any node so it can be retrieved without passing a target dom reference (like normal variables)
-
-Attribute({name: 'thisFileVar1', value: (n:number)=>Math.PI*n})
-const thisFileVar1 = Attribute({name: 'thisFileVar1'})
+ *
+ * ```ts
+ * // setting a type or node's attribute value
+ * Attribute<SomeInterface>({name: 'id', value: 'org.foo.Super'})
+ * Attribute({target: anImpl1, name: 'class', value: 'logger'})
+ * Attribute({target: aNode, name: 'otherNode', value: aFunctionDeclaration})
+ *
+ * // getting (previously set) attribute values
+ * const v = Attribute({name: 'id'})
+ * console.log(thisFileVar1==='org.foo.Super')
+ *
+ * // if target is not passed at all the the variable won't be bound to any node so it can be retrieved without passing a target dom reference (like normal variables)
+ * Attribute({name: 'thisFileVar1', value: (n:number)=>Math.PI*n})
+ * const thisFileVar1 = Attribute({name: 'thisFileVar1'})
+ * console.log(thisFileVar1(8))
 ```
  */
 export function Attribute<T = any>(config: AttributeOptions, t?: any): (string | Stats)[] {
@@ -55,6 +56,7 @@ export class AttributeClass extends AbstractExtractor {
     const config = (this.getOptionsFromFistArg(n) as AttributeOptions) || {}
     let output: string | NodeWithInfo | undefined
     if (config.value || config.action === 'set') {
+      // SETTER
       output =
         typeof config.value === 'undefined'
           ? 'undefined'
@@ -62,7 +64,6 @@ export class AttributeClass extends AbstractExtractor {
               info: config.name,
               node: config.value
             } as NodeWithInfo)
-
       const argument = variableAccessor(
         { name: config.name },
         index,
@@ -77,6 +78,8 @@ export class AttributeClass extends AbstractExtractor {
     } else if (config.action === 'list') {
       throw 'not impl'
     } else {
+      // GETTER
+      // HEADS UP: we need tp pass a function predicate as a string in order to find our attribute since it was set by another call. TODO: provide better API in the core
       return {
         argument:
           variableAccessor(
@@ -89,7 +92,6 @@ export class AttributeClass extends AbstractExtractor {
         prependToFile: '""'
       }
     }
-    // return this.buildExtractorResult(n, output!, getter, index, options, config || {})
   }
 
   protected parseOptionValue(name: string, value: Node | undefined): any {